Update Next version, re-organize directory structure
