+++ package.json
@@ -11,9 +11,9 @@
   "dependencies": {
     "axios": "^0.26.1",
     "mongoose": "^6.3.0",
-    "next": "12.1.5",
-    "react": "18.0.0",
-    "react-dom": "18.0.0",
+    "next": "^12.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
