+++ src/utils/parseGame.ts
@@ -1,4 +1,4 @@
-import {Game, IMove, ITag} from "../../../lib/models/game"
+import {Game, IMove, ITag} from "../lib/models/game"
 
 /**
  * @param data Array of .pgn file strings from api/handleUpload/apiRoute.ts
